requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-restframework==0.0.1
django-sslserver==0.22
django-timezone==0.3
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
